Feature/plugins cleanup

+++ compute/kubernetes/resources/pvc.go
@@ -5,8 +5,8 @@ import (
 	"context"
 	"fmt"
 	"html/template"
-	"os"
 
+	"github.com/ohsu-comp-bio/funnel/config"
 	"github.com/ohsu-comp-bio/funnel/logger"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -16,14 +16,12 @@ import (
 
 // Create the Worker/Executor PVC from config/kubernetes-pvc.yaml
 // TODO: Move this config file to Helm Charts so users can see/customize it
-func CreatePVC(taskId string, namespace string, bucket string, region string, tplFile string, client kubernetes.Interface, log *logger.Logger) error {
-	tpl, err := os.ReadFile(tplFile)
-	if err != nil {
-		return fmt.Errorf("reading template: %v", err)
-	}
+func CreatePVC(taskId string, config *config.Config, client kubernetes.Interface, log *logger.Logger) error {
+
+	jobNamespace := config.Kubernetes.JobsNamespace
 
 	// Load templates
-	t, err := template.New(taskId).Parse(string(tpl))
+	t, err := template.New(taskId).Parse(config.Kubernetes.PVCTemplate)
 	if err != nil {
 		return fmt.Errorf("parsing template: %v", err)
 	}
@@ -32,9 +30,9 @@ func CreatePVC(taskId string, namespace string, bucket string, region string, tp
 	var buf bytes.Buffer
 	err = t.Execute(&buf, map[string]interface{}{
 		"TaskId":    taskId,
-		"Namespace": namespace,
-		"Bucket":    bucket,
-		"Region":    region,
+		"Namespace": jobNamespace,
+		"Bucket":    config.GenericS3[0].Bucket,
+		"Region":    config.GenericS3[0].Region,
 	})
 	if err != nil {
 		return fmt.Errorf("%v", err)
@@ -51,7 +49,7 @@ func CreatePVC(taskId string, namespace string, bucket string, region string, tp
 		return fmt.Errorf("failed to decode PVC spec")
 	}
 
-	_, err = client.CoreV1().PersistentVolumeClaims(namespace).Create(context.Background(), pvc, metav1.CreateOptions{})
+	_, err = client.CoreV1().PersistentVolumeClaims(jobNamespace).Create(context.Background(), pvc, metav1.CreateOptions{})
 	if err != nil {
 		return fmt.Errorf("%v", err)
 	}
@@ -62,10 +60,13 @@ func CreatePVC(taskId string, namespace string, bucket string, region string, tp
 // Add this helper function for PVC cleanup
 func DeletePVC(ctx context.Context, taskID string, namespace string, client kubernetes.Interface, log *logger.Logger) error {
 	name := fmt.Sprintf("funnel-worker-pvc-%s", taskID)
-	err := client.CoreV1().PersistentVolumeClaims(namespace).Delete(ctx, name, metav1.DeleteOptions{})
-
-	if err != nil {
-		return fmt.Errorf("deleting shared PVC: %v", err)
+	_, err := client.CoreV1().PersistentVolumeClaims(namespace).Get(ctx, name, metav1.GetOptions{})
+	if err == nil {
+		log.Debug("deleting Worker PVC", "taskID", taskID)
+		err := client.CoreV1().PersistentVolumeClaims(namespace).Delete(ctx, name, metav1.DeleteOptions{})
+		if err != nil {
+			return fmt.Errorf("deleting shared PVC: %v", err)
+		}
 	}
 
 	return nil
